LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/urange-parsing-expected.txt


FAIL "u/**/+/**/a/**/?" => "U+A0-AF" assert_equals: expected "U+A0-AF" but got ""
FAIL "u+abc" => "U+ABC" assert_equals: expected "U+ABC" but got ""
FAIL "U+abc" => "U+ABC" assert_equals: expected "U+ABC" but got ""
FAIL "u+ABC" => "U+ABC" assert_equals: expected "U+ABC" but got ""
FAIL "U+ABC" => "U+ABC" assert_equals: expected "U+ABC" but got ""
FAIL "u+AbC" => "U+ABC" assert_equals: expected "U+ABC" but got ""
FAIL "u+efg" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+ abc" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u +abc" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u + abc" is invalid assert_equals: expected "U+1357" but got ""
FAIL "U + a b c" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+a" => "U+A" assert_equals: expected "U+A" but got ""
FAIL "u+aa" => "U+AA" assert_equals: expected "U+AA" but got ""
FAIL "u+aaa" => "U+AAA" assert_equals: expected "U+AAA" but got ""
FAIL "u+aaaa" => "U+AAAA" assert_equals: expected "U+AAAA" but got ""
FAIL "u+aaaaa" => "U+AAAAA" assert_equals: expected "U+AAAAA" but got ""
FAIL "u+aaaaaaa" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+a?" => "U+A0-AF" assert_equals: expected "U+A0-AF" but got ""
FAIL "u+a??" => "U+A00-AFF" assert_equals: expected "U+A00-AFF" but got ""
FAIL "u+a???" => "U+A000-AFFF" assert_equals: expected "U+A000-AFFF" but got ""
FAIL "u+a????" => "U+A0000-AFFFF" assert_equals: expected "U+A0000-AFFFF" but got ""
FAIL "u+aaaaaa?" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+aaaaa??" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+aaaa???" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+aaa????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+aa?????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+a??????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+a?a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+aaaaaa" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+a?????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u/**/+0a/**/?" => "U+A0-AF" assert_equals: expected "U+A0-AF" but got ""
FAIL "u+0a" => "U+A" assert_equals: expected "U+A" but got ""
FAIL "U+0a0" => "U+A0" assert_equals: expected "U+A0" but got ""
FAIL "u+0aaaaa" => "U+AAAAA" assert_equals: expected "U+AAAAA" but got ""
FAIL "u+0aaaaaa" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0a0000" => "U+A0000" assert_equals: expected "U+A0000" but got ""
FAIL "u+0a00000" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0aaaaa0" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+00000a" => "U+A" assert_equals: expected "U+A" but got ""
FAIL "u+00000aa" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+00000a0" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+000000a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0a????" => "U+A0000-AFFFF" assert_equals: expected "U+A0000-AFFFF" but got ""
FAIL "u+0a?????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+00a????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+22222a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+1e9a" => "U+1E9A" assert_equals: expected "U+1E9A" but got ""
FAIL "u/**/+0/**/?" => "U+0-F" assert_equals: expected "U+0-F" but got ""
FAIL "u/**/0" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+00" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+000" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+0000" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+00000" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+000000" => "U+0" assert_equals: expected "U+0" but got ""
FAIL "u+0000000" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+00000?" => "U+0-F" assert_equals: expected "U+0-F" but got ""
FAIL "u+0?????" => "U+0-FFFFF" assert_equals: expected "U+0-FFFFF" but got ""
FAIL "u+0?a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+000000?" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+00000??" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0??????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+1e3" => "U+1E3" assert_equals: expected "U+1E3" but got ""
FAIL "u+1e-20" => "U+1E-20" assert_equals: expected "U+1E-20" but got ""
FAIL "u+222222" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+2?????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u/**/+0/**/-0a" => "U+0-A" assert_equals: expected "U+0-A" but got ""
FAIL "u+0-0a" => "U+0-A" assert_equals: expected "U+0-A" but got ""
FAIL "u+000000-0aaaaa" => "U+0-AAAAA" assert_equals: expected "U+0-AAAAA" but got ""
FAIL "u+0000000-0a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0-0aaaaaa" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0-000000a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0+0a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0?-0a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0-0a?" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+222222-22222a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u/**/+0/**/-1" => "U+0-1" assert_equals: expected "U+0-1" but got ""
FAIL "u+0-1" => "U+0-1" assert_equals: expected "U+0-1" but got ""
FAIL "u-0-1" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u-0+1" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0+1" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+000000-000001" => "U+0-1" assert_equals: expected "U+0-1" but got ""
FAIL "u+0000000-1" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0-0000001" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+0-222222" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u/**/+/**/?" => "U+0-F" assert_equals: expected "U+0-F" but got ""
FAIL "u+?" => "U+0-F" assert_equals: expected "U+0-F" but got ""
FAIL "u+?????" => "u+0-FFFFF" assert_equals: expected "U+0-FFFFF" but got ""
FAIL "u+???????" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+?a" is invalid assert_equals: expected "U+1357" but got ""
FAIL "u+??????" is invalid assert_equals: expected "U+1357" but got ""
FAIL u+a is a valid selector assert_equals: expected "u + a" but got ".error"
